+++ src/metaflac/operations.c
@@ -565,7 +565,7 @@ void write_metadata(const char *filename, FLAC__StreamMetadata *block, unsigned 
 			PPR; printf("  sample_rate: %u Hz\n", block->data.stream_info.sample_rate);
 			PPR; printf("  channels: %u\n", block->data.stream_info.channels);
 			PPR; printf("  bits-per-sample: %u\n", block->data.stream_info.bits_per_sample);
-			PPR; printf("  total samples: %" PRIu64 "\n", (uint64_t)block->data.stream_info.total_samples);
+			PPR; printf("  total samples: %" PRIu64 "\n", block->data.stream_info.total_samples);
 			PPR; printf("  MD5 signature: ");
 			for(i = 0; i < 16; i++) {
 				printf("%02x", (unsigned)block->data.stream_info.md5sum[i]);
@@ -592,7 +592,7 @@ void write_metadata(const char *filename, FLAC__StreamMetadata *block, unsigned 
 			PPR; printf("  seek points: %u\n", block->data.seek_table.num_points);
 			for(i = 0; i < block->data.seek_table.num_points; i++) {
 				if(block->data.seek_table.points[i].sample_number != FLAC__STREAM_METADATA_SEEKPOINT_PLACEHOLDER) {
-					PPR; printf("    point %u: sample_number=%" PRIu64 ", stream_offset=%" PRIu64 ", frame_samples=%u\n", i, (uint64_t)block->data.seek_table.points[i].sample_number, (uint64_t)block->data.seek_table.points[i].stream_offset, block->data.seek_table.points[i].frame_samples);
+					PPR; printf("    point %u: sample_number=%" PRIu64 ", stream_offset=%" PRIu64 ", frame_samples=%u\n", i, block->data.seek_table.points[i].sample_number, block->data.seek_table.points[i].stream_offset, block->data.seek_table.points[i].frame_samples);
 				}
 				else {
 					PPR; printf("    point %u: PLACEHOLDER\n", i);
@@ -610,7 +610,7 @@ void write_metadata(const char *filename, FLAC__StreamMetadata *block, unsigned 
 			break;
 		case FLAC__METADATA_TYPE_CUESHEET:
 			PPR; printf("  media catalog number: %s\n", block->data.cue_sheet.media_catalog_number);
-			PPR; printf("  lead-in: %" PRIu64 "\n", (uint64_t)block->data.cue_sheet.lead_in);
+			PPR; printf("  lead-in: %" PRIu64 "\n", block->data.cue_sheet.lead_in);
 			PPR; printf("  is CD: %s\n", block->data.cue_sheet.is_cd? "true":"false");
 			PPR; printf("  number of tracks: %u\n", block->data.cue_sheet.num_tracks);
 			for(i = 0; i < block->data.cue_sheet.num_tracks; i++) {
@@ -618,7 +618,7 @@ void write_metadata(const char *filename, FLAC__StreamMetadata *block, unsigned 
 				const FLAC__bool is_last = (i == block->data.cue_sheet.num_tracks-1);
 				const FLAC__bool is_leadout = is_last && track->num_indices == 0;
 				PPR; printf("    track[%u]\n", i);
-				PPR; printf("      offset: %" PRIu64 "\n", (uint64_t)track->offset);
+				PPR; printf("      offset: %" PRIu64 "\n", track->offset);
 				if(is_last) {
 					PPR; printf("      number: %u (%s)\n", (unsigned)track->number, is_leadout? "LEAD-OUT" : "INVALID");
 				}
@@ -633,7 +633,7 @@ void write_metadata(const char *filename, FLAC__StreamMetadata *block, unsigned 
 					for(j = 0; j < track->num_indices; j++) {
 						const FLAC__StreamMetadata_CueSheet_Index *index = track->indices+j;
 						PPR; printf("        index[%u]\n", j);
-						PPR; printf("          offset: %" PRIu64 "\n", (uint64_t)index->offset);
+						PPR; printf("          offset: %" PRIu64 "\n", index->offset);
 						PPR; printf("          number: %u\n", (unsigned)index->number);
 					}
 				}
